[WIP] More work towards first working version.
The Spaceport is now fully fleshed out and waits for new connections,
adding them to the list of satellites.

GroundControl is now also fully fleshed out and listens to satellites for
new events.  These get passed into a mutex-locked queue and a condition
variable is triggered to wake up the switches.

New Switch object started to route the messages to the satellites that
have registered for them.